requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic[email]
google-generativeai